Add layer grid
(has to be improved, but it's working)

+++ src/Util/view.cpp
@@ -2,6 +2,6 @@
 
 void moveViewTowardsPoint(sf::View& view, float x, float y, float speed)
 {
-    view.setCenter(view.getCenter().x + (x - view.getCenter().x) * speed,
-                   view.getCenter().y + (y - view.getCenter().y) * speed);
+    view.setCenter((int)(view.getCenter().x + (x - view.getCenter().x) * speed),
+                   (int)(view.getCenter().y + (y - view.getCenter().y) * speed));
 }
